Refactor code

+++ app/controller/user_controller.py
@@ -2,30 +2,28 @@
 User controller
 """
 
-from flask_jwt_extended import jwt_required
-from flask_restplus import Resource, Namespace, fields
-from app.service.user_service import get_all_user, add_user, get_user, update_user, delete_user
-from app.model.user import User
+from flask_jwt_extended import jwt_required, get_jwt_identity
+from flask_restplus import Resource, Namespace, fields, abort
 
+from app.controller.struct_controller import FEATURE_COLLECTION_MODEL, structures_to_geojson
+from app.model.user import User
+from app.service.user_service import \
+    get_all_user, add_user, get_user, update_user, delete_user, \
+    get_favourites_by_user, get_all_structure_by_user
 
 API = Namespace('User', description='User related operations', path='/users')
 
 USER_MODEL = API.model('user', {
     'id': fields.Integer(required=False, description='User identifier'),
     'username': fields.String(required=True, description='Username of the user'),
-    'password_hash': fields.String(required=False, description='User password hash'),
     'created_on': fields.DateTime(required=False, description='User creation date')
 })
 
-USER_CREATE = API.model('user', {
+USER_PASSWORD_MODEL = API.model('user_password', {
     'username': fields.String(required=True, description='Username of the user'),
     'password': fields.String(required=True, description='User password'),
 })
 
-USER_UPDATE = API.model('user', {
-    'username': fields.String(required=True, description='Username of the user'),
-    'password': fields.String(required=False, description='User password'),
-})
 
 @API.route('')
 class UserListController(Resource):
@@ -39,15 +37,14 @@ class UserListController(Resource):
         """List all users"""
         return get_all_user()
 
-
     @API.doc('create_user', security=None)
-    @API.expect(USER_CREATE)
-    @API.marshal_with(USER_CREATE, code=201)
+    @API.expect(USER_PASSWORD_MODEL)
+    @API.marshal_with(USER_MODEL, code=201)
     def post(self):
         """Add a new user"""
         user = User(username=API.payload['username'], password=API.payload['password'])
         add_user(user)
-        return API.payload, 201
+        return user, 201
 
 
 @API.route('/<int:user_id>')
@@ -64,11 +61,10 @@ class UserController(Resource):
         """Get the user"""
         return get_user(user_id)
 
-
     @jwt_required
     @API.doc('update_user')
-    @API.expect(USER_UPDATE)
-    @API.marshal_with(USER_UPDATE)
+    @API.expect(USER_PASSWORD_MODEL)
+    @API.marshal_with(USER_MODEL)
     def put(self, user_id):
         """Update the user"""
         user = get_user(user_id)
@@ -81,11 +77,41 @@ class UserController(Resource):
         update_user(updated_user)
         return updated_user
 
-
     @jwt_required
     @API.doc('delete_user')
     @API.response(204, 'User deleted')
+    @API.response(401, 'Cannot delete user. An user can only delete itself')
     def delete(self, user_id):
         """Delete a user given its identifier"""
-        delete_user(user_id)
-        return '', 204
+        if user_id == get_jwt_identity():
+            delete_user(user_id)
+            return '', 204
+
+        abort(401)
+        return
+
+
+@API.route('/<int:user_id>/structures')
+class StructureUserController(Resource):
+    """
+    Show a list of all structure from a user
+    """
+
+    @API.doc('list_resources_from_user', security=None)
+    @API.marshal_with(FEATURE_COLLECTION_MODEL)
+    def get(self, user_id):
+        """List all structures of an user"""
+        return structures_to_geojson(get_all_structure_by_user(user_id))
+
+
+@API.route('/<int:user_id>/favorites')
+class FavoritesUserController(Resource):
+    """
+    Show a list of all structure from a user
+    """
+
+    @API.doc('list_favourites_of_user', security=None)
+    @API.marshal_with(FEATURE_COLLECTION_MODEL)
+    def get(self, user_id):
+        """List all favourites of an user"""
+        return structures_to_geojson(get_favourites_by_user(user_id))
